requirements.txt
tensorflow>=2.9.0
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0
